+++ include/aikido/control/JacobianExecutor.hpp
@@ -105,6 +105,10 @@ using JacobianVelocityExecutor = JacobianExecutor<ExecutorType::VELOCITY>;
 using JacobianEffortExecutor = JacobianExecutor<ExecutorType::EFFORT>;
 // Note: No position, as jacobian should only be used with derivatives.
 
+// Define constexprs (required for linking)
+template <ExecutorType T> constexpr size_t JacobianExecutor<T>::SE3_SIZE;
+template <ExecutorType T> constexpr double JacobianExecutor<T>::DEFAULT_LAMBDA;
+
 } // namespace control
 } // namespace aikido