+++ include/aikido/robot/util.hpp
@@ -196,6 +196,25 @@ inline std::set<std::string> dofNamesFromSkeleton(
   return ret;
 }
 
+/// Make a pre-made exceptional std::future
+template <typename T>
+inline std::future<T> make_exceptional_future(std::string error_message)
+{
+  auto promise = std::promise<T>();
+  promise.set_exception(
+      std::make_exception_ptr(std::runtime_error(error_message)));
+  return promise.get_future();
+}
+
+/// Make a pre-made ready std::future
+template <typename T>
+inline std::future<T> make_ready_future(T obj)
+{
+  auto promise = std::promise<T>();
+  promise.set_value(obj);
+  return promise.get_future();
+}
+
 } // namespace util
 } // namespace robot
 } // namespace aikido